requirements.txt
tqdm
Pillow
argparse
psycopg[binary]
transformers